requirements.txt
flask
python-docx
gunicorn
pandas
